+++ app/services/activity/runnings/contracts/create_contract.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 module Activity
-  module Running
+  module Runnings
     module Contracts
-      class UpdateContract < Dry::Validation::Contract
+      class CreateContract < Dry::Validation::Contract
         schema(Schemas::RunningSchema)
       end
     end
