+++ app/services/activity/runnings/operations/container.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Activity
-  module Running
+  module Runnings
     module Operations
       class Container
         extend Dry::Container::Mixin
