+++ app/services/activity/runnings/transactions/create_transaction.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
 module Activity
-  module Running
+  module Runnings
     module Transactions
-      class UpdateTransaction < BaseTransaction
+      class CreateTransaction < BaseTransaction
         step :validate, with: 'shared.validate'
+        step :build, with: 'shared.build'
         step :assign_attributes, with: 'assign_attributes'
         step :persist, with: 'shared.persist'
       end
