+++ app/services/activity/runnings/transactions/update_transaction.rb
@@ -1,11 +1,10 @@
 # frozen_string_literal: true
 
 module Activity
-  module Running
+  module Runnings
     module Transactions
-      class CreateTransaction < BaseTransaction
+      class UpdateTransaction < BaseTransaction
         step :validate, with: 'shared.validate'
-        step :build, with: 'shared.build'
         step :assign_attributes, with: 'assign_attributes'
         step :persist, with: 'shared.persist'
       end
