+++ spec/services/activity/runnings/operations/container_spec.rb
@@ -3,10 +3,10 @@
 require 'rails_helper'
 require 'dry/transaction/operation'
 
-RSpec.describe Activity::Running::Operations::Container do
+RSpec.describe Activity::Runnings::Operations::Container do
   subject(:container) { described_class }
 
   it 'has assign_attributes operation' do
-    expect(container['assign_attributes']).to be_instance_of(Activity::Running::Operations::AssignAttributes)
+    expect(container['assign_attributes']).to be_instance_of(Activity::Runnings::Operations::AssignAttributes)
   end
 end
